Feature/add aoe to projectile (#11)
* projectile now hits only one target

* specify rust version in toml

* initial spawn of unanimated explosion

* explosion spawn an animates somewhere on the screen

* explosions spawning by attaching component to enemies

* explosions are now triggered via events

* adjusted replease notes

+++ src/game/projectiles/systems.rs
@@ -1,31 +1,24 @@
-use crate::game::player::components::Score;
-
-use super::constants::DAMAGE;
-use super::{
-    super::{
-        enemies::{components::Enemy, constants::SPRITE_DIAMETER as ENEMY_SPRITE_DIAMETER},
-        player::components::{CastTimer, Health, Player},
-    },
-    constants::{
-        EXPLOSION_TEXTURE_COLUMNS, EXPLOSION_TEXTURE_PATH, EXPLOSION_TEXTURE_ROWS,
-        EXPLOSION_TEXTURE_SIZE,
-    },
+use super::super::{
+    enemies::{components::Enemy, constants::SPRITE_DIAMETER as ENEMY_SPRITE_DIAMETER},
+    explosion::events::Explode,
+    player::components::{CastTimer, Health, Player},
 };
+use super::constants::DAMAGE;
 use super::{
     components::Projectile,
     constants::{PROJECTILE_TEXTURE_PATH, SPEED, SPRITE_DEPTH, SPRITE_DIAMETER},
 };
 use bevy::{
-    asset::{AssetServer, Assets},
+    asset::AssetServer,
+    ecs::event::EventWriter,
     math::{
         bounding::{BoundingCircle, IntersectsVolume},
-        UVec2, Vec2, Vec3,
+        Vec2, Vec3,
     },
     prelude::{Commands, Entity, Query, Res, Transform, With, Without},
     sprite::Sprite,
     time::Time,
 };
-use std::collections::HashSet;
 
 pub fn spawn(
     mut player_query: Query<(&mut CastTimer, &Transform), With<Player>>,
@@ -76,13 +69,15 @@ pub fn movement(mut projectile_query: Query<(&mut Projectile, &mut Transform)>, 
 pub fn hit_target(
     projectile_query: Query<(&Projectile, &Transform, Entity)>,
     mut enemy_query: Query<(&Transform, &mut Health), With<Enemy>>,
+    mut explode_events: EventWriter<Explode>,
     mut commands: Commands,
-    mut player_query: Query<&mut Score, With<Player>>,
 ) {
-    let mut entities_to_despawn = HashSet::new();
     for (projectile, transform, entity) in projectile_query.iter() {
-        let mut entity_exists = true;
-        for (enemy_transform, mut health) in enemy_query.iter_mut() {
+        if projectile.is_finished() {
+            commands.entity(entity).despawn();
+            continue;
+        }
+        for (enemy_transform, mut enemy_health) in enemy_query.iter_mut() {
             let projectile_collider =
                 BoundingCircle::new(transform.translation.truncate(), SPRITE_DIAMETER / 2.);
             let enemy_collider = BoundingCircle::new(
@@ -90,20 +85,14 @@ pub fn hit_target(
                 ENEMY_SPRITE_DIAMETER / 2.,
             );
             if projectile_collider.intersects(&enemy_collider) {
-                entities_to_despawn.insert(entity);
-                health.deal_damage(DAMAGE);
-                entity_exists = false;
-                if let Ok(mut score) = player_query.get_single_mut() {
-                    score.score += 1;
-                }
+                commands.entity(entity).despawn();
+                enemy_health.deal_damage(DAMAGE);
+                explode_events.send(Explode {
+                    transform: enemy_transform.clone(),
+                });
+                break;
             }
         }
-        if projectile.is_finished() && entity_exists {
-            entities_to_despawn.insert(entity);
-        }
-    }
-    for entity in entities_to_despawn {
-        commands.entity(entity).despawn();
     }
 }
 
@@ -112,49 +101,3 @@ pub fn despawn(projectile_query: Query<Entity, With<Projectile>>, mut commands: 
         commands.entity(entity).despawn();
     }
 }
-
-// pub fn spawn_explosion(
-//     asset_server: Res<AssetServer>,
-//     mut texture_atlas_layouts: ResMut<Assets<TextureAtlasLayout>>,
-//     mut commands: Commands,
-// ) {
-//     // load the sprite sheet using the `AssetServer`
-//     let texture = asset_server.load(EXPLOSION_TEXTURE_PATH);
-
-//     // the sprite sheet has 7 sprites arranged in a row, and they are all 24px x 24px
-//     let layout = TextureAtlasLayout::from_grid(
-//         UVec2::splat(EXPLOSION_TEXTURE_SIZE),
-//         EXPLOSION_TEXTURE_COLUMNS,
-//         EXPLOSION_TEXTURE_ROWS,
-//         None,
-//         None,
-//     );
-//     let texture_atlas_layout = texture_atlas_layouts.add(layout);
-
-//     let sprite = Sprite {
-
-//         custom_size: Some(Vec2::new(SPRITE_DIAMETER, SPRITE_DIAMETER)),
-//         ..Default::default()
-//     };
-
-//     let sprite_bundle = SpriteBundle {
-//         sprite: sprite,
-//         texture: asset_server.load(TEXTURE_PATH),
-//         transform: Transform::from_xyz(0., 0., 0.),
-//         ..Default::default()
-//     };
-
-//     commands.spawn((
-//         Sprite {
-//             image: texture.clone(),
-//             texture_atlas: Some(TextureAtlas {
-//                 layout: texture_atlas_layout.clone(),
-//                 index: animation_config_1.first_sprite_index,
-//             }),
-//             ..default()
-//         },
-//         Transform::from_scale(Vec3::splat(6.0)).with_translation(Vec3::new(-50.0, 0.0, 0.0)),
-//         LeftSprite,
-//         animation_config_1,
-//     ));
-// }
